add tool
